+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,0 +1,11 @@
+package lotto.domain.lotto;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+
+class LottoTest {
+
+}
